blockchain_contracts: Fix clippy error
